update potential issue in two_to_three
when face_rotation is true, we remove new_tet1 and new_tet2. But some face of new_tet0 could be attached to one of them. In that case there will be a corresponding face of new_tet0 which it's attached to, updated two_to_three to detach then attach to the correct face of new_tet0.